requirements.txt
h3==3.7.2
RPi.gpio==0.7.0
colorzero==2.0
gpiozero==1.6.2
protobuf==3.15.6
dbus-python==1.2.16
nmcli==0.5.0
sentry-sdk==1.1.0
pycairo==1.20.1
PyGObject==3.40.1
retry==0.9.2
requests==2.26.0
hm-hardware-defs==0.1.5
python-dotenv